[scheduling]: handle synchronization of windowless
