+++ src/examples/arm.cpp
@@ -42,7 +42,7 @@ int main()
 
     robot_dart::RobotDARTSimu simu;
 #ifdef GRAPHIC
-    simu.set_graphics(std::make_shared<robot_dart::gui::magnum::Graphics<>>(&simu));
+    simu.set_graphics(std::make_shared<robot_dart::gui::magnum::Graphics>(&simu));
 #endif
     simu.add_descriptor(std::make_shared<StateDesc>(&simu));
     simu.add_robot(g_robot);
