+++ src/examples/control_loop.cpp
@@ -16,7 +16,7 @@ int main()
 
     robot_dart::RobotDARTSimu simu(1e-3);
 #ifdef GRAPHIC
-    simu.set_graphics(std::make_shared<robot_dart::gui::magnum::Graphics<>>(&simu));
+    simu.set_graphics(std::make_shared<robot_dart::gui::magnum::Graphics>(&simu));
 #endif
 
     simu.add_robot(robot);
