+++ src/examples/dof_maps.cpp
@@ -25,7 +25,7 @@ int main()
     float dt = 0.001;
     robot_dart::RobotDARTSimu simu(dt);
 #ifdef GRAPHIC
-    auto graphics = std::make_shared<robot_dart::gui::magnum::Graphics<>>(&simu);
+    auto graphics = std::make_shared<robot_dart::gui::magnum::Graphics>(&simu);
     simu.set_graphics(graphics);
     graphics->look_at({0., 3.5, 2.}, {0., 0., 0.25});
 #endif
