+++ src/examples/pendulum.cpp
@@ -53,7 +53,7 @@ int main()
 
     robot_dart::RobotDARTSimu simu;
 #ifdef GRAPHIC
-    simu.set_graphics(std::make_shared<robot_dart::gui::magnum::Graphics<>>(&simu));
+    simu.set_graphics(std::make_shared<robot_dart::gui::magnum::Graphics>(&simu));
 #endif
     // <Type>(desc_period)
     simu.add_descriptor<StateDesc>(2);
